+++ packages/app/src/views/list/custom.ts
@@ -1,0 +1,12 @@
+export class Custom {
+  folders = [
+    { name: 'Photos', icon: 'folder', addDate: 'Jan 9, 2015' },
+    { name: 'Recipes', icon: 'folder', addDate: 'Jan 17, 2015' },
+    { name: 'Work', icon: 'folder', addDate: 'Jan 28, 2015' }
+  ];
+
+  files = [
+    { name: 'Vacation Itinerary', icon: 'insert_drive_file', addDate: 'Jan 10, 2015' },
+    { name: 'Kitchen Remodel', icon: 'insert_drive_file', addDate: 'Jan 20, 2015' }
+  ];
+}